SEVERAL MAJOR CHANGES: Reformatted combine_sn_and_host, read_binned_arrays, save_binned_templates_as arrays files so that they all use the same CombineSnAndHost or BinTemplate classes and so that there is no code repition. Also removing the continuum of the host galaxy and then multiplying by the continuum so that the spectral features are weighted by the continuum.
Continuum is now divided instead of subtracted when removing it because this is the more correct way to do this.

Correcting the way that outerVal (region outside the wavelength range with data) is set to 0.5 instead of 0 before training and viewing.

Using **kwarg arguments for the arrays in create_arrays.py so that filenames and typeNames aren't mandatory arguments (while images and labels still are mandatory).

Changed the snCoeff (SN coefficients) used when adding SN + host during training. Was previously incorrectly training on more host than SN (defined as a list in create_arrays.py).

Moved oversampling (and augmenting with noise) of arrays to deep_learning_multilayer.py instead of saving this in the npy.gz training set arrays in order to save hardisk memory and to fix compression and decompression issues due to memory errors.